Make the auth cache hash keys part of the public API.
This allows API consumers to make better use of the data returned by
functions like svn_config_walk_auth_data(), eliminates duplicate
definitions of some of these keys, and makes it easier to add new
key definitions in the future.

There is no reason for these definitions to be private since we are
unlikely to remove existing hash keys used by the on-disk format.

* subversion/include/svn_config.h
  (SVN_CONFIG_AUTHN_USERNAME_KEY,
   SVN_CONFIG_AUTHN_PASSWORD_KEY,
   SVN_CONFIG_AUTHN_PASSPHRASE_KEY,
   SVN_CONFIG_AUTHN_PASSTYPE_KEY,
   SVN_CONFIG_AUTHN_ASCII_CERT_KEY,
   SVN_CONFIG_AUTHN_FAILURES_KEY): Declare and document. Equivalent to
    local AUTHN_* macro definitions in the various files below.

* subversion/libsvn_subr/simple_providers.c,
  subversion/libsvn_subr/ssl_client_cert_pw_providers.c,
  subversion/libsvn_subr/ssl_server_trust_providers.c,
  subversion/libsvn_subr/username_providers.c,
  subversion/svn/auth-cmd.c: Remove local definitions of hash keys and
   use the global definitions instead.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1572640 13f79535-47bb-0310-9956-ffa450edef68