Fix compiler errors regarding index values for membuffer cache.
The maximum entries in a membuffer cache is limited to the max value of an
apr_uint32_t.  But does pointer arithmatic to calculate indexes into the
cache.  On 64-bit platforms the pointers are going to be 64-bits wide, so
the calculation will be 64-bits wide.  The results will stay within the
range of the apr_uint32_t so it's safe to just cast these calculations.

* subversion/libsvn_subr/cache-membuffer.c
  (free_spare_group): Cast to apr_uint32_t when calculating first_spare_group.
  (drop_entry): Cast to apr_uint32_t when calculating last_in_group.
  (find_entry): Cast to apr_uint32_t when calculating previous and next on
    group headers.




git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1574128 13f79535-47bb-0310-9956-ffa450edef68