Use the correct type for revision numbers.  Similar to r1575019 to fsfs.
While yes we're now limiting them to 32-bit effectively (despite the
svn_revnum_t type being wider on many platforms today), we shouldn't be using
a different type anywhere in our code so that we can change the type and not
have to go looking for all the places where we've done things like this.  Yes,
this uses more memory at runtime, but this is the long term consequence of our
type choice for revision numbers.

* subversion/libsvn_fs_x/fs.h
  (representation_cache_key_t): Use svn_revnum_t instead of apr_uint32_t.



git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1575376 13f79535-47bb-0310-9956-ffa450edef68