FSFS: Add new config option to pack after every commit for debug purposes.
This option is deliberately not in our default config file because we don't
want to encourage people to use this for normal operation.  This option will
ultimately replace the -DPACK_AFTER_EVERY_COMMIT knob used for debugging
packing.

To Subversion Admins that find this commit: You want to continue using
post-commit hooks to trigger packing.  This option blocks commits from
showing as completed to users until the pack completes.  Which is probably
not desireable.

* subversion/libsvn_fs_fs/fs.h
  (CONFIG_SECTION_DEBUG, CONFIG_OPTION_PACK_AFTER_COMMIT): New.
  (fs_fs_data_t): Add pack_after_commit member.

* subversion/libsvn_fs_fs/fs_fs.c
  (read_config): Read the new option disabling it for formats that don't
    support packing.

* subversion/libsvn_fs_fs/tree.c
  (svn_fs_fs__commit_txn): Pack if configured to do so.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1577362 13f79535-47bb-0310-9956-ffa450edef68