[Reverted in r1589518].
Fix a scalability issue with FSFS repositories: The lock mutexes are
part of the process-wide per-repo shared structs.  Using only the
UUID as identification for the repo is not enough in case a repo gets
duplicated or the admins don't follow good practice.  The result may
be serializing all commits globally across all repos.

Now, we simply add the normalized repo path to the hash key but keep
the UUID as well.  The latter allows us to detect and handle replaced
repositories gracefully in some cases.

* subversion/libsvn_fs_fs/fs.c
  (fs_serialized_init): Add FS path to the hash lookup key.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1589262 13f79535-47bb-0310-9956-ffa450edef68