Support connection to gpg-agent's that are using the standard socket location.
This is now the default behavior of gpg-agent (though many distributions
change that at build time).  This means that GPG_AGENT_INFO is not required
if the gpg-agent's socket is in the standard location.  The gpg-agent man page
describes how a client should determine the socket to connect to.

Some distributions of GPG rely on this and avoid setting GPG_AGENT_INFO
altogether, making our implementation require extra setup that GPG itself
doesn't.  For example the GPGTools distribution of GPG on the Mac.

* subversion/libsvn_subr/gpg_agent.c
  (find_running_gpg_agent): If there is no GPG_AGENT_INFO environment variable
    try the standard location.  Add some documentation of the GPG_AGENT_INFO
    environment variable format to make it easier to understand what the code
    is doing.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1600368 13f79535-47bb-0310-9956-ffa450edef68