Fix another problem in the gen_win_dependencies.py script.  There are some cases when we want to print a warning about an unresolved dependency, e.g.: [[[   WARNING: 'opensslv.h' not found   Use '--with-openssl' to configure openssl location. ]]]
Prior to this changeset, we would instead end up with printing a stacktrace
("... global name 'show_warning' is not defined") instead of issuing a
warning and continuing to work.  The reason for that is somewhat trivial --
we inconsistently use the names 'show_warning' and 'show_warnings' across
the script and sometimes encounter a situation when the name of the function
argument does not match the name of the variable in its body.  Fix this by
using a consistent naming scheme across the whole script file.

* build/generator/gen_win_dependencies.py
  (_find_httpd, _find_openssl, _find_jdk, _find_swig, _find_serf,
   _find_sasl, _find_libintl): Use 'show_warnings' instead of 'show_warning'
    as the argument names.  The discrepancies are only in the openssl,
    jdk, and swig-related functions, but using a consistent naming scheme
    *everywhere* could probably avoid the same mistakes in the future.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1604039 13f79535-47bb-0310-9956-ffa450edef68