Tweak the svn_delta_editor debug editor, that prints diagnostic messages for debugging.
Add the 'abort' method, which was missing -- it didn't even forward the call
to the real, wrapped editor. Add an optional prefix, which is useful to
distinguish the interleaved messages from nested editors. Tweak the "DBG:"
prefix to include a space, "DBG: ", for compatibility with SVN_DBG(). Print
to stdout instead of stderr, for compatibility with SVN_DBG() and with our
test suite which ignores stdout lines starting with "DBG:".

* subversion/libsvn_delta/debug_editor.c
  (edit_baton): Add a 'prefix' member.
  (write_indent): Write the prefix instead of just "DBG:".
  (abort_edit): New function.
  (svn_delta__get_debug_editor): Don't bother getting a 'default' editor,
    since we want to override all the methods. Use stdout instead of stderr.
    Initialize the prefix.

* subversion/libsvn_delta/debug_editor.h
  (svn_delta__get_debug_editor): Add the 'prefix' argument. Document the
    changes.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1614313 13f79535-47bb-0310-9956-ffa450edef68