Following up on r1619418, don't use not-present, excluded and server excluded node kinds for reporting a tree conflict. These nodes are NOT in the working copy, so they are not part of the creation of a tree conflict.
Whether deleted nodes should or shouldn't be handled as nodes in the WC
really depends on the specific kind of tree conflict. In many merge cases they
shouldn't, while for update/switch they usually should.

Note that the original check of local_kind == svn_node_unknown would almost
never be true, as this would only be valid for actual only nodes, which
by definition are deleted/absent in the working copy.

* subversion/libsvn_wc/conflicts.c
  (read_tree_conflict_desc): Don't pass TRUE for show_hidden unless you
    explicitly want to handle these cases. Don't look at deleted nodes
    when merging, as those nodes can be completely replaced by the merge.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1619717 13f79535-47bb-0310-9956-ffa450edef68